Rename and restructure benchmark folder
