+++ py-polars/tests/benchmark/groupby-datagen.R
